chatbot-server/requirements.txt
fastapi
uvicorn
weaviate-client
pydantic
python-dotenv
